pyproject.toml
[project]
name = "aws13th-social-backend-gpffh20"
version = "0.1.0"
description = "Backend API for Cloud Community Service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "pydantic[email]>=2.0.0,<3.0.0",
    "bcrypt>=4.0.0,<5.0.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.14.11",
]
